+++ targets/__init__.py
@@ -5,4 +5,4 @@ defineOption('fileEncodingDecider', ExtensionBasedFileEncodingDecider({
 	'.json':'utf-8',
 	'.xml':'utf-8',
 	'.yaml':'utf-8', '.yml':'utf-8',
-}, default='shift-jis'))
+}, default='ascii'))